RavenDB-27020 Use ES2020 syntax in ace editor
